+++ src/core/ScriptPer.gd
@@ -11,7 +11,11 @@ func _init(per_msg: perMessage).(
 		per_msg.trigger_object) -> void:
 	# The name of the type of per we're seeking gets its own var
 	script_name = per_msg.per_seek
-	var ret = _find_subjects(per_msg.subjects)
+	if get_property(SP.KEY_ORIGINAL_PREVIOUS):
+		prev_subjects = per_msg.prev_subjects
+	else:
+		prev_subjects = per_msg.subjects
+	var ret = _find_subjects()
 	if ret is GDScriptFunctionState: # Still working.
 		ret = yield(ret, "completed")
 	# We emit a signal when done so that our ScriptingEngine